wTUJdlU/requirements.txt
django
djangorestframework==3.14.0

# for the swagger ui
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1

# for generating the random data
faker

# for CORS (allows the frontend to use the api)
django-cors-headers==3.14.0
